requirements.txt
albumentations==1.3.1
datasets==2.13.1
deepspeed==0.10.0
huggingface-hub==0.15.1
langcodes==3.3.0
matplotlib==3.7.2
numpy==1.25.0
opencv-python==4.7.0.72
pandas==2.0.2
peft @ git+https://github.com/huggingface/peft.git@13e53fc7ee5d89d59b16523051006dddf0fb7a49
Pillow==9.5.0
prompt-toolkit==3.0.38
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.97
tensorboard==2.14.0
timm==0.4.12
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
transformers==4.28.1